+++ course/02-EventEmitter/CHANGELOG.md
@@ -1,7 +1,13 @@
 # @udemy-nodejs/02-eventemitter
 
+## 1.0.2
+
+### Patch Changes
+
+- chore: 将示例文件 example1.js 和 example2.js 整合为一个文件
+
 ## 1.0.1
 
-### Major Changes
+### Patch Changes
 
 - feat: 添加了 node.js 的 events 库的使用,并且理解手写事件发射器,添加了示例